BUG/MINOR: ssl: Possible memleak when allowing the 0RTT data buffer.
â
As the server early data buffer is allocated in the middle of the loop
used to allocate the SSL session without being freed before retrying,
this leads to a memory leak.
â
To fix this we move the section of code responsible of this early data buffer
alloction after the one reponsible of allocating the SSL session.
â
Must be backported to 2.1 and 2.0.

(cherry picked from commit 3139c1b198bbcc14c6940f214234afd004110387)
Signed-off-by: Willy Tarreau <w@1wt.eu>