MINOR: lua: Add a function to get a reference on a table in the stack
The hlua_checktable() function may now be used to create and return a
reference on a table in stack, given its position. This function ensures it
is really a table and throws an exception if not.

This patch is mandatory to allow the support of the filters written in lua.

(cherry picked from commit ba9e21dc68044e2ce5ffce317e8786f1079f6fee)
[cf: Required for another backport]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>