BUG/MINOR: lua: Ignore the reserve to know if a channel is full or not
The Lua function Channel.is_full() should not take care of the reserve because
it is not called from a producer (an applet for instance). From an action, it is
allowed to overwrite the buffer reserve.

This patch should be backported as far as 1.7. But it must be adapted for 1.8
and lower because there is no HTX on these versions.

(cherry picked from commit 0ec740eaee5b919b7d568daf251c8ce218143639)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 99471972da9d7516165adb8c104771ef8219fbb1)
Signed-off-by: Willy Tarreau <w@1wt.eu>
(cherry picked from commit 0136ebba371d598c5c44f4e41ec1229337616379)
Signed-off-by: Willy Tarreau <w@1wt.eu>